Introduced two new classes: LocalPod and Sandbox.
The Sandbox represents the entire contents of the POD_ROOT (normally SOURCE_ROOT/Pods). A LocalPod represents a pod that has been installed within the Sandbox.

These two classes can be used as better homes for various pieces of logic currently spread throughout the installation process and provide a better API for working with the contents of this directory.

So far I've just refactored the downloaders to use them.