My earlier Resolver/TargetInstaller refactoring broke the handling of target-specific specs, because previously each TargetInstaller would resolve it's specs separately.
Instead, we resolve everything in one go as before, but have the resolver return a hash of
target dependencies to resolved specifications, so when the target installers are run, we only
install the pods for that target that should be in there.

Likewise, we should only call the post-install hook of a specification with the target installers
it belongs to.