Immediately print output from Executable when in verbose mode.
Fixes #276.

I understand that the reason was to indent the output, but I think it's
more important to have instant feedback.

I went on a yak shaving, but now we can capture output and show it at
the same time. This should allow for better error checking. The problem
is that, for instance, Git (almost) always exits with non-zero exit
statusses, so that's not a good way to check it. We should probably
check the actual output.

This is all related to #266.